feat: start to rework memory accessing
